AuraScribe_Backend/requirements.txt
# AuraScribe Backend Dependencies
# Flask and extensions
Flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
flask-limiter==3.5.0
Werkzeug==2.3.7

# Environment and utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Google Cloud / AI
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0
google-adk

# Deepgram Speech-to-Text
deepgram-sdk==3.1.0

# Production server
gunicorn==21.2.0
eventlet==0.33.3

# Security
python-jose[cryptography]==3.3.0

# Redis for session storage
redis==5.0.1